pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer is a website that analyzes the specified pages for SEO suitability."
authors = ["Dmitry Grigoriev <diman03051995@mail.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
validators = "^0.34.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
